validation and handle errors

+++ controllers/books.js
@@ -3,19 +3,28 @@ const mongodb = require('../data/database');
 const getAllBooks = async (req, res) => {
     //#swagger.tags=['Books']
     const result = await mongodb.getDatabase().db().collection('books').find();
-    result.toArray().then((books) => {
+    result.toArray((err, books) => {
+        if (err) {
+            res.status(400).json({message: err});
+        }
         res.setHeader('Content-Type', 'application/json');
-        res.status(200).json(books);
+        res.status(200).json(books)
     });
 };
 
 const getSingle = async(req, res) => {
     //#swagger.tags=['Books']
+    if (!req.params.id) {
+        return res.status(400).json('Must use a valid ISBN to find a book.');
+    }
     const bookId = req.params.id;
     const result = await mongodb.getDatabase().db().collection('books').find({_id: bookId});
-    result.toArray().then((books) => {
+    result.toArray((err, books) => {
+        if (err) {
+            res.status(400).json({message: err});
+        }
         res.setHeader('Content-Type', 'application/json');
-        res.status(200).json(books);
+        res.status(200).json(books)
     });
 }
 
@@ -33,7 +42,7 @@ const insertBook = async (req, res) => {
     const result = await mongodb.getDatabase().db().collection('books').insertOne(book);
 
     if (result.acknowledged) {
-      res.status(201).json({ message: "Book created successfully" });
+      res.status(201).json("Book created successfully");
     } else {
       res.status(500).json("Failed to insert book");
     }
@@ -46,6 +55,9 @@ const insertBook = async (req, res) => {
 
 const updateBook = async(req, res) => {
     //#swagger.tags=['Books']
+    if (!req.params.id) {
+        return res.status(400).json('Must use a valid ISBN to find a book.');
+    }
     const bookId = req.params.id;
     const book = {
         title: req.body.title,
@@ -55,7 +67,7 @@ const updateBook = async(req, res) => {
     };
     const result = await mongodb.getDatabase().db().collection('books').replaceOne({_id: bookId}, book);
     if (result.modifiedCount > 0){
-        res.status(201).json({ message: "Book updated successfully" });
+        res.status(201).json("Book updated successfully");
     } else {
         res.status(500).json(result.error || 'Some error occurred while updating the contact.');
     }
@@ -63,10 +75,13 @@ const updateBook = async(req, res) => {
 
 const deleteBook = async(req, res) => {
     //#swagger.tags=['Books']
+    if (!req.params.id) {
+        return res.status(400).json('Must use a valid ISBN to find a book.');
+    }
     const bookId = req.params.id;
     const result = await mongodb.getDatabase().db().collection('books').deleteOne({_id: bookId});
     if (result.deletedCount > 0) {
-        res.status(201).json({ message: "Book deleted successfully" });
+        res.status(201).json("Book deleted successfully");
     } else {
         res.status(500).json(result.error || 'Book not found');
     }
